split files out to headers and source
