+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "*.pl": "prolog",
+        "iostream": "cpp"
+    }
+}
