+++ src/cpp/director.cpp
@@ -1,0 +1,11 @@
+#include <iostream>
+#include <vector>
+
+#include "../hpp/events.hpp"
+
+class Director {
+    public:
+        Event dummyEvent() {
+            return Event();
+        };
+};
