+++ src/hpp/director.hpp
@@ -1,0 +1,10 @@
+#ifndef DIRECTOR_HPP_
+#define DIRECTOR_HPP_
+
+#include "../hpp/events.hpp"
+
+class Director {
+    public:
+        Event dummyEvent();
+};
+#endif // DIRECTOR_HPP_
