+++ src/hpp/main.hpp
@@ -1,0 +1,6 @@
+#ifndef MAIN_HPP_
+#define MAIN_HPP_
+
+int main();
+
+#endif // MAIN_HPP_
